requirements-dev.txt
# Development and testing dependencies
coverage==7.3.2
flake8==6.1.0
black==23.9.1
isort==5.12.0
bandit==1.7.5
safety==2.3.5
pytest==7.4.3
pytest-django==4.7.0